refactor(core-kernel): move calculateForgingInfo into core kernel utils
