+++ packages/core-kernel/src/contracts/shared/forging-info.ts
@@ -1,0 +1,6 @@
+export interface ForgingInfo {
+    currentForger: number;
+    nextForger: number;
+    blockTimestamp: number;
+    canForge: boolean;
+}
