+++ packages/core-kernel/src/contracts/shared/index.ts
@@ -1,3 +1,4 @@
 export * from "./rounds";
 export * from "./download-block";
 export * from "./dynamic-fee";
+export * from "./forging-info";
