+++ packages/core-kernel/src/utils/index.ts
@@ -1,3 +1,4 @@
+import { calculateForgingInfo } from "./calculate-forging-info";
 import { calculateApproval, calculateForgedTotal } from "./delegate-calculator";
 import { calculateLockExpirationStatus, calculateTransactionExpiration } from "./expiration-calculator";
 import { formatTimestamp } from "./format-timestamp";
@@ -15,5 +16,6 @@ export const delegateCalculator = { calculateApproval, calculateForgedTotal };
 export const expirationCalculator = { calculateTransactionExpiration, calculateLockExpirationStatus };
 export const roundCalculator = { calculateRound, isNewRound };
 export const supplyCalculator = { calculate };
+export const forgingInfoCalculator = { calculateForgingInfo };
 
 export { formatTimestamp, isBlockChained, getBlockNotChainedErrorMessage, isWhitelisted, Plugins };
